[migration] LoginState.tsx 변환 #156
